+++ app/store/middleware/__tests__/tracking.spec.js
@@ -1,7 +1,7 @@
 import simple from 'simple-mock';
 
-import AppConstants from '../../constants/AppConstants';
-import trackingMiddleware, { track } from './tracking';
+import AppConstants from '../../../constants/AppConstants';
+import trackingMiddleware, { track } from '../tracking';
 
 describe('store/middleware/tracking', () => {
   let dispatch;
